Currently implementing changes to SdlGameView.cpp and SdlGameView.h to use OpenGL functionality and drawing in an SDL window.
CMake now includes paths to Glew and OpenGL libraries which is required for OpenGL with SDL implementation.

I cleaned up and encapsulated a lot of the SdlGameView code into smaller functions.  There is now an init() function to initialize the game window as well as an initGL() function which will initialize OpenGL features.  An exit() function will perform cleanup on the SDL window and OpenGL context, then close the game.

A drawSquare() function will use OpenGL commands to draw a simple square.  Using GL textures, rotations, translations, and scaling we can use this instance square to create many of the objects we need in the game (namely vehicles and bases).

DO NOT MERGE THIS BRANCH TO MASTER YET! The OpenGL implementation is still not finished. Though the initializations are in working order, OpenGL won't draw anything except for it's clear screen buffer (a white background).  Still need to investigate the cause of this and finish drawing a vehicle with OpenGL commands.

+++ Vehicle.cpp
@@ -4,7 +4,7 @@
 #include "Vehicle.h"
 
 //constructor
-Vehicle::Vehicle(int cX, int cY, int end, int ang, double vel){
+Vehicle::Vehicle(int cX, int cY, int end, double ang, double vel){
     //values will be passed to GameState
     //values are placeholders for testing purposes
     velocity = vel;        //current velocity
